Added systems container, added basic file system.

+++ source/main.cpp
@@ -1,8 +1,10 @@
 #include "game.hpp"
 
+#include "op_result.hpp"
+
 int main()
 {
   blocks_game::Game game;
-
-  return game.Run();
+  
+  return (int)game.Run();
 }
